move tallying to CellStage
