+++ bluemira/radiation_transport/neutronics/make_pre_cell.py
@@ -21,8 +21,10 @@ from bluemira.geometry.coordinates import (
     get_bisection_line,
 )
 from bluemira.geometry.error import GeometryError
+from bluemira.geometry.plane import calculate_plane_dir
 from bluemira.geometry.solid import BluemiraSolid
 from bluemira.geometry.tools import (
+    get_wire_plane_intersect,
     is_convex,
     make_polygon,
     raise_error_if_overlap,
@@ -890,3 +892,131 @@ class DivertorPreCellArray:
             copy of itself
         """
         return DivertorPreCellArray(self.pre_cells.copy())
+
+
+class PreCellStage:
+    """Stage of making pre-cells"""
+
+    def __init__(self, blanket: PreCellArray, divertor: DivertorPreCellArray):
+        """Check convexity after initialization"""
+        self.blanket = blanket.copy()
+        self.divertor = divertor
+        # 1. stretch first blanket cell in PreCellArray to reach div_start_wire
+        div_start_wire = self.divertor[0].cw_wall.restore_to_wire()
+        # pull everything down to: div_start_wire.
+        # Alternatively, choose div_start_wire=self.divertor[0].outline
+        old_vv_wire = self.blanket[0].vv_wire
+        ext_pt, i_high, i_low = np.insert(self.blanket[0].vertex, 1, 0, axis=0).T[:3]
+        i_end = get_wire_plane_intersect(
+            div_start_wire, *calculate_plane_dir(i_high, i_low)
+        )
+        # v_end = get_wire_plane_intersect(
+        #     div_start_wire,
+        #     *calculate_plane_dir(old_vv_wire.end_point().xyz.flatten(),
+        #     old_vv_wire.start_point().xyz.flatten())
+        # )
+        in_wire = make_polygon(np.array([i_high, i_end]).T, closed=False)
+        vv_wire = make_polygon(
+            np.array([
+                self.divertor[0].vv_wire.end_point,
+                old_vv_wire.end_point().xyz.flatten(),
+            ]).T,
+            closed=False,
+        )
+        ex_wire = make_polygon(
+            np.array([self.divertor[0].vertex.T[0], ext_pt]).T, closed=False
+        )
+        new_start_cell = PreCell(in_wire, vv_wire, ex_wire)
+        self.blanket[0] = new_start_cell
+
+        # 2. stretch first blanket cell in PreCellArray to reach div_end_wire
+        div_end_wire = self.divertor[-1].ccw_wall.restore_to_wire()
+        old_vv_wire = self.blanket[-1].vv_wire
+        i_low, i_high, ext_pt = np.insert(self.blanket[-1].vertex, 1, 0, axis=0).T[-3:]
+        i_start = get_wire_plane_intersect(
+            div_end_wire, *calculate_plane_dir(i_high, i_low)
+        )
+        # v_end = get_wire_plane_intersect(
+        #     div_end_wire,
+        #     *calculate_plane_dir(old_vv_wire.start_point().xyz.flatten(),
+        #     old_vv_wire.end_point().xyz.flatten())
+        # )
+        in_wire = make_polygon(np.array([i_start, i_high]).T, closed=False)
+        vv_wire = make_polygon(
+            np.array([
+                old_vv_wire.start_point().xyz.flatten(),
+                self.divertor[-1].vv_wire.start_point,
+            ]).T,
+            closed=False,
+        )
+        ex_wire = make_polygon(
+            np.array([ext_pt, self.divertor[-1].vertex.T[-1]]).T, closed=False
+        )
+        new_end_cell = PreCell(in_wire, vv_wire, ex_wire)
+        self.blanket[-1] = new_end_cell
+
+        # re-initialize so that the cell_walls are re-calculated
+        self.blanket = self.blanket.copy()
+
+    def external_coordinates(self) -> npt.NDArray:
+        """
+        Get the outermost coordinates of the tokamak cross-section from pre-cell array
+        and divertor pre-cell array.
+        Runs clockwise, beginning at the inboard blanket-divertor joint.
+
+        Returns
+        -------
+        :
+            All vertices on the exterior of the blanket and the divertor.
+        """
+        return np.concatenate([
+            self.blanket.exterior_vertices(),
+            self.divertor.exterior_vertices()[::-1],
+        ])
+
+    def bounding_box(self) -> tuple[float, ...]:
+        """
+        Get bounding box of pre cell stage
+
+        Returns
+        -------
+        z_max:
+            The maximum height of the bounding box.
+        z_min:
+            The minimum height of the bounding box.
+        r_max:
+            The maximum major radius reached by the pre-cells.
+        -r_max:
+            The minimum major radius reached by the pre-cells. Due to axial symmetry,
+            this must be =-r_max.
+        """
+        all_ext_vertices = self.external_coordinates()
+        z_min = all_ext_vertices[:, -1].min()
+        z_max = all_ext_vertices[:, -1].max()
+        r_max = max(abs(all_ext_vertices[:, 0]))
+        return z_max, z_min, r_max, -r_max
+
+    def half_bounding_box(self) -> tuple[float, ...]:
+        """
+        Get bounding box of the 2D poloidal cross-section of the right-hand half of the
+        reactor.
+
+        Returns
+        -------
+        z_max:
+            The maximum height of the bounding box.
+        z_min:
+            The minimum height of the bounding box.
+        r_max:
+            The maximum major radius reached by the pre-cells.
+        r_min:
+            The minimum major radius reached by the pre-cells on one side of the xz cross
+            section. This is typically non-zero because pre-cells should not cross the
+            z-axis of symmetry.
+        """
+        all_ext_vertices = self.external_coordinates()
+        z_min = all_ext_vertices[:, -1].min()
+        z_max = all_ext_vertices[:, -1].max()
+        r_max = max(abs(all_ext_vertices[:, 0]))
+        r_min = min(abs(all_ext_vertices[:, 0]))
+        return z_max, z_min, r_max, r_min
